add user profile photo and name in config page modify change on size of the screen
